Updating change_variant and change_animation
No longer forcing the current klass and variant for the right side of promotion.

For now, I will force the left side to use the unit's default variant. This can go either way, and I'm willing to change it based on feedback, but the feature needs to exist first so I can actually get good feedback.

Also fixing a bug since "None" is actually a valid variant type in most cases.