CRIT_ADD refactored, THRACIA_CRIT added
I approached this problem through the equations widget because I think it's the simplest solution

I edited combat_calcs because that's what the damage component calls, and also deals with very particular edge cases.

Thracia_Crit is setup to not be required in the code so people pulling this change won't have to add it in.